+++ tasmota/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Change redesigning ESP8266 GPIO internal representation in line with ESP32
 - Change new IR Raw compact format (#9444)
 - Change MAX31865 driver to support up to 6 thermocouples selected by ``MX31865 CS`` instead of ``SSPI CS`` (#9103)
+- Change A4988 optional microstep pin selection
 - Add optional support for Mitsubishi Electric HVAC by David Gwynne (#9237)
 - Add optional support for Orno WE517-Modbus energy meter by Maxime Vincent (#9353)